Update types Use other activities
