+++ package.json
@@ -20,10 +20,10 @@
     "vue": "^2.6.10",
     "vue-class-component": "^7.1.0",
     "vue-i18n": "^8.15.0",
-    "vue-notification": "^1.3.16",
+    "vue-notification": "^1.3.19",
     "vue-property-decorator": "^8.3.0",
     "vue-router": "^3.1.3",
-    "vuetify": "^2.1.6",
+    "vuetify": "^2.1.7",
     "vuex": "^3.1.1",
     "vuex-class-component": "^2.0.4",
     "vuex-persist": "^2.1.1"
