+++ src/modules/AniList/types/AniListActivityType.ts
@@ -1,0 +1,7 @@
+export enum AniListActivityType {
+  TEXT = 'TEXT',
+  ANIME_LIST = 'ANIME_LIST',
+  MANGA_LIST = 'MANGA_LIST',
+  MESSAGE = 'MESSAGE',
+  MEDIA_LIST = 'MEDIA_LIST',
+}
