+++ src/modules/AniList/types/AniListListStatus.ts
@@ -1,0 +1,8 @@
+export enum AniListListStatus {
+  CURRENT = 'CURRENT',
+  PLANNING = 'PLANNING',
+  COMPLETED = 'COMPLETED',
+  DROPPED = 'DROPPED',
+  PAUSED = 'PAUSED',
+  REPEATING = 'REPEATING',
+}
