+++ src/modules/AniList/types/AniListScoreFormat.ts
@@ -1,0 +1,7 @@
+export enum AniListScoreFormat {
+  POINT_100 = 'POINT_100',
+  POINT_10_DECIMAL = 'POINT_10_DECIMAL',
+  POINT_10 = 'POINT_10',
+  POINT_5 = 'POINT_5',
+  POINT_3 = 'POINT_3',
+}
