+++ src/modules/AniList/types/AniListSeason.ts
@@ -1,0 +1,6 @@
+export enum AniListSeason {
+  WINTER = 'WINTER',
+  SPRING = 'SPRING',
+  SUMMER = 'SUMMER',
+  FALL = 'FALL',
+}
