+++ src/modules/AniList/types/AniListType.ts
@@ -1,0 +1,4 @@
+export enum AniListType {
+  ANIME = 'ANIME',
+  MANGA = 'MANGA',
+}
