+++ src/modules/AniList/types/IAniListCoverImage.ts
@@ -1,0 +1,14 @@
+export interface IAniListCoverImage {
+  /**
+   * @var {string} medium contains the medium size thumbnail link
+   */
+  medium: string;
+  /**
+   * @var {string} large contains the large size thumbnail link
+   */
+  large: string;
+  /**
+   * @var {string} extraLarge contains the extraLarge size thumbnail link
+   */
+  extraLarge: string;
+}
