+++ src/modules/AniList/types/IAniListList.ts
@@ -1,0 +1,13 @@
+import { IAniListEntry } from './IAniListEntry';
+import { AniListListStatus } from './AniListListStatus';
+
+export interface IAniListList {
+  /**
+   * @var {IAniListEntry[]} entries contains the list's user entries
+   */
+  entries: IAniListEntry[];
+  /**
+   * @var {AniListListStatus} status contains the list's status
+   */
+  status: AniListListStatus;
+}
