+++ src/modules/AniList/types/IAniListMediaDate.ts
@@ -1,0 +1,14 @@
+export interface IAniListMediaDate {
+  /**
+   * @var {number} year contains the date's year
+   */
+  year: number;
+  /**
+   * @var {number} month contains the date's month
+   */
+  month: number;
+  /**
+   * @var {number} day contains the date's day
+   */
+  day: number;
+}
