+++ src/modules/AniList/types/IAniListMediaListCollection.ts
@@ -1,0 +1,8 @@
+import { IAniListList } from './IAniListList';
+
+export interface IAniListMediaListCollection {
+  /**
+   * @var {IAniListList[]} lists contains the user's lists
+   */
+  lists: IAniListList[];
+}
