+++ src/modules/AniList/types/IAniListMediaStudio.ts
@@ -1,0 +1,8 @@
+import { IAniListMediaStudioNode } from './IAniListMediaStudioNode';
+
+export interface IAniListMediaStudio {
+  /**
+   * @var {IAniListMediaStudioNode[]} nodes contains the informations about the media's studios
+   */
+  nodes: IAniListMediaStudioNode[];
+}
