+++ src/modules/AniList/types/IAniListMediaStudioNode.ts
@@ -1,0 +1,10 @@
+export interface IAniListMediaStudioNode {
+  /**
+   * @var {string} name contains the studio's name
+   */
+  name: string;
+  /**
+   * @var {boolean} isAnimationStudio
+   */
+  isAnimationStudio: boolean;
+}
