+++ src/modules/AniList/types/IAniListSeasonPreview.ts
@@ -1,0 +1,7 @@
+import { AniListSeason, IAniListSeasonPreviewMedia } from '.';
+
+export interface IAniListSeasonPreview {
+  season: AniListSeason;
+  seasonYear: number;
+  media: IAniListSeasonPreviewMedia[];
+}
