+++ src/modules/AniList/types/IAniListUserAvatar.ts
@@ -1,0 +1,10 @@
+export interface IAniListUserAvatar {
+  /**
+   * @var {string} medium contains the URL to the user's avatar (medium-size)
+   */
+  medium: string;
+  /**
+   * @var {string} large contains the URL to the user's avatar (large-size)
+   */
+  large: string;
+}
