+++ src/modules/AniList/types/IAniListUserMediaListOptions.ts
@@ -1,0 +1,6 @@
+export interface IAniListUserMediaListOptions {
+  /**
+   * @var {string} scoreFormat contains the user's preferred scoring system format
+   */
+  scoreFormat: string;
+}
