+++ src/modules/AniList/types/IAniListUserOptions.ts
@@ -1,0 +1,3 @@
+export interface IAniListUserOptions {
+  displayAdultContent: boolean;
+}
