+++ src/modules/AniList/types/IAniListUserStats.ts
@@ -1,0 +1,6 @@
+export interface IAniListUserStats {
+  /**
+   * @var {number} watchedTime contains the amount of minutes the user has watched anime
+   */
+  watchedTime: number;
+}
