+++ src/plugins/AniList/responses/LatestActivitiesResponse.ts
@@ -1,0 +1,7 @@
+import { IAniListActivity } from '@/modules/AniList/types';
+
+export interface LatestActivitiesResponse<T = IAniListActivity> {
+  page: {
+    activities: T[];
+  };
+}
