+++ src/plugins/AniList/responses/index.ts
@@ -1,4 +1,5 @@
 export { BaseResponse } from './BaseResponse';
+export { LatestActivitiesResponse } from './LatestActivitiesResponse';
 export { MediaResponse } from './MediaResponse';
 export { SearchResponse } from './SearchResponse';
 export { SeasonPreviewResponse } from './SeasonPreviewResponse';
